Add PhraseProvider
