+++ HW2/Valeriy_Satkov/CreateDoorsAndLevels/CreateDoorsAndLevels/Interfaces/IPhraseProvider.cs
@@ -1,0 +1,11 @@
+﻿namespace CreateDoorsAndLevels.Interfaces
+{
+    /* Associates a key phrase with sentences in the source (File, DB, ...)
+     */
+    interface IPhraseProvider
+    {
+        /* Return text from source by phraseKey
+         */
+        string GetPhrase(string phraseKey);
+    }
+}
