+++ HW2/Valeriy_Satkov/CreateDoorsAndLevels/CreateDoorsAndLevels/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
+</packages>
